% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{assign_numeric_palettes}
\alias{assign_numeric_palettes}
\title{Assign a continuous palette identifier to each numeric variable.}
\usage{
assign_numeric_palettes(num_vars, pool = continuous_palette_pool())
}
\arguments{
\item{num_vars}{character vector of numeric metadata variable names.}

\item{pool}{palette pool from [continuous_palette_pool()].}
}
\value{
named list of `"package::palette"` strings.
}
\description{
Cycles through `pool` if there are more numeric variables than palettes.
}
\keyword{internal}
