% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{continuous_palette_ids}
\alias{continuous_palette_ids}
\title{All `package::palette` identifiers accepted in `__numeric__` CSV rows.}
\usage{
continuous_palette_ids()
}
\description{
All `package::palette` identifiers accepted in `__numeric__` CSV rows.
}
\keyword{internal}
