% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_helpers.R
\name{continuous_palette_pool}
\alias{continuous_palette_pool}
\title{Continuous palette pool used for numeric metadata variables.}
\usage{
continuous_palette_pool()
}
\value{
data.frame with columns `package`, `palette` (subset of
  `paletteer::palettes_c_names`).
}
\description{
Continuous palette pool used for numeric metadata variables.
}
\keyword{internal}
