% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyloseq_extended_graphical_methods.R
\name{fast_tax_glom}
\alias{fast_tax_glom}
\title{Fast alternative to [tax_glom()] from phyloseq.extended package.}
\usage{
fast_tax_glom(
  physeq,
  taxrank = rank_names(physeq)[1],
  bad_empty = c("", " ", "\\t")
)
}
\arguments{
\item{physeq}{Required. \code{\link{phyloseq-class}} object}

\item{taxrank}{A character string specifying the taxonomic level that you want to agglomerate over. Should be among the results of \code{rank_names(physeq)}. The default value is \code{rank_names(physeq)[1]}, which may agglomerate too broadly for a given experiment. You are strongly encouraged to try different values for this argument.}

\item{bad_empty}{(Optional). Character vector. Default to empty white spaces and tabs. Defines the bad/empty values that should be replaced by "Unknown".}
}
\value{
A taxonomically-agglomerated \code{\link{phyloseq-class}} object.
}
\description{
Fast alternative to [tax_glom()] from phyloseq.extended package.
}
\details{
fast_tax_glom differs from [tax_glom()] in three important ways:
* It is based on dplyr function and thus much faster
* It does not preserve the phy_tree slot
* It only preserves taxonomic ranks up to \code{taxrank} and discard all other ones.
The "archetype" (OTU name) and sequence for each group are chosen as those from the most abundant taxa
within that group (for compatibility with [tax_glom()])
}
\seealso{
[tax_glom()], [merge_taxa()]
}
