% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_alpha.R
\name{mod_alpha_ui}
\alias{mod_alpha_ui}
\alias{mod_alpha_server}
\title{mod_alpha_ui and mod_alpha_server}
\usage{
mod_alpha_ui(id)

mod_alpha_server(input, output, session, r = r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
