% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_compo.R
\name{mod_compo_ui}
\alias{mod_compo_ui}
\alias{mod_compo_server}
\title{mod_compo_ui and mod_compo_server}
\usage{
mod_compo_ui(id)

mod_compo_server(input, output, session, r = r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
