% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_diffanalysis.R
\name{mod_diffanalysis_ui}
\alias{mod_diffanalysis_ui}
\alias{mod_diffanalysis_server}
\title{mod_diffanalysis_ui and mod_diffanalysis_server}
\usage{
mod_diffanalysis_ui(id)

mod_diffanalysis_server(input, output, session, r = r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
