% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_diffexplore.R
\name{mod_diffexplore_ui}
\alias{mod_diffexplore_ui}
\alias{mod_diffexplore_server}
\title{mod_diffexplore_ui and mod_diffexplore_server}
\usage{
mod_diffexplore_ui(id)

mod_diffexplore_server(input, output, session, r = r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
