% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_metadata_subset.R
\name{mod_metadata_subset_ui}
\alias{mod_metadata_subset_ui}
\alias{mod_metadata_subset_server}
\title{mod_metadata_subset_ui and mod_metadata_subset_server}
\usage{
mod_metadata_subset_ui(id)

mod_metadata_subset_server(input, output, session, r = r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
