% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_taxaboxplot.R
\name{mod_taxaboxplot_ui}
\alias{mod_taxaboxplot_ui}
\alias{mod_taxaboxplot_server}
\title{mod_taxaboxplot_ui and mod_taxaboxplot_server}
\usage{
mod_taxaboxplot_ui(id)

mod_taxaboxplot_server(input, output, session, r = r)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
A shiny Module.
}
\keyword{internal}
