% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bars_fun.R
\name{bars_fun}
\alias{bars_fun}
\title{Barplots plotly}
\usage{
bars_fun(
  data,
  rank = "Genus",
  top = 10,
  Ord1 = NULL,
  sample_labels = FALSE,
  split = FALSE,
  split_sid_order = FALSE,
  relative = TRUE,
  autoorder = TRUE,
  ylab = "Abundance",
  outfile = "plot_compo.html",
  verbose = TRUE,
  pal = NULL
)
}
\arguments{
\item{data}{a phyloseq object}

\item{rank}{Taxonomy rank to merge features (among rank_names(data), or
'ASV' for no glom)}

\item{top}{Number of top taxa to plot}

\item{Ord1}{Variable used to order samples (X axis) or split the barplot}

\item{sample_labels}{If TRUE, x axis labels are sample IDs}

\item{split}{If TRUE, make a facet_wrap-like plot grouped by Ord1}

\item{split_sid_order}{If TRUE, keep sample IDs order from metadata in split}

\item{relative}{Plot relative (TRUE) or raw abundance (FALSE)}

\item{autoorder}{Automatic ordering of x axis labels with mixedsort}

\item{ylab}{Y axis title}

\item{outfile}{Output html file (NULL to skip saving)}

\item{verbose}{Print log messages}

\item{pal}{A color palette}
}
\value{
An interactive plotly barplot
}
\description{
Barplots plotly
}
